third_party/blink/web_tests/virtual/pna-navigations-warning/external/wpt/fetch/private-network-access/window-open-existing.tentative.https.window_include=from-private-expected.txt

This is a testharness.js-based test.
[FAIL] private to local: failed preflight.
  assert_equals: expected "timeout" but got "success"
[FAIL] private to local: missing CORS headers.
  assert_equals: expected "timeout" but got "success"
[FAIL] private to local: missing PNA header.
  assert_equals: expected "timeout" but got "success"
Harness: the test ran to completion.
